更新 main.cpp
修改了your_lover_age的数据类型(由短整型修改为浮点数)

+++ main.cpp
@@ -44,7 +44,7 @@ int main(){
     }else{
         your_gender=="男" ? wife_or_husband="老婆" : wife_or_husband="老公";
     }
-    short your_lover_age=0;
+    float your_lover_age=0;
     string get_age;
     do{
         cout<<"你"<<wife_or_husband<<"的年龄:";
